Sync: 69. Sqrt(x) (Easy) - Leet2Git

+++ 69_Sqrtx.cpp
@@ -1,12 +1,12 @@
 class Solution {
 public:
     int mySqrt(int x) {
-        int low = 0;
-        int high = x;
-        int ans = 0;
+        long long low = 0;
+        long long high = x;
+        long long ans = 0;
 
         while(low <= high) {
-            int mid = low + (high - low) / 2;
+            long long mid = low + (high - low) / 2;
 
             if(mid * mid <= x) {
                 ans = mid;
@@ -17,6 +17,6 @@ public:
             }
         }
 
-        return ans;
+        return (int)ans;
     }
 };
